requirements.txt
Flask==2.3.3
gunicorn==20.1.0
requests==2.31.0
python-dotenv==1.0.0
cloudipsp==1.0.4